fix name issues
