+++ features/jera_list_grid/__brick__/{{name.snakeCase()}}/components/{{name.snakeCase()}}_item_view.dart
