+++ features/jera_list_grid/__brick__/{{name.snakeCase()}}/di/{{name.snakeCase()}}_module.dart
@@ -1,7 +1,7 @@
 import '../../../support/utils/service_locator/app_module.dart';
 import '../../../support/utils/service_locator/service_locator.dart';
-import '../{{name.snakeCase()}}_route_view.dart';
-import '../{{name.snakeCase()}}_route_view_model.dart';
+import '../{{name.snakeCase()}}_view.dart';
+import '../{{name.snakeCase()}}_view_model.dart';
 
 class {{name.pascalCase()}}Module extends AppModule {
   static const String route = '/{{name.snakeCase()}}_route';
